requirements.txt
openai>=1.0.0
pyserial>=3.5
pillow>=10.1.0
pydantic>=2.0.0
python-dotenv>=1.0.0
fastapi>=0.100.0
uvicorn>=0.23.0
requests>=2.31.0
-e ./niimprint
